Use last RTT to extend timeout
If the configured timeout is a multiple of the interval, and there is packet
loss, then it can happen that the reply to the last request arrives after the
timeout triggers. This leads to superfluous state changes ONLINE -> OFFLINE ->
ONLINE within the same second. By postponing the timeout a bit (and relative to
the last RTT) we give the last reply a chance to arrive before the timeout
triggers and we can avoid that.