Fix build with kernel 5.15 and musl
Currently, with kernel 5.15 headers and musl building is failing with
redefinition errors due to a conflict between the kernel and musl
headers.

Musl is able to suppress the conflicting kernel header definitions
if they are included after the standard libc ones, however the reformat
in commit 8cee8ac moved these kernel definitions before <netinet/in.h>
and this suppression has stopped working and is breaking the builds.

Solve this by moving the <netinet/in.h> include before the ICMP kernel
include in the icmp.c

Fixes: 8cee8ac5bece ("Add curly braces and reformat")